requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
langchain==0.1.14
openai==1.12.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose==3.3.0
tenacity==8.2.3
sqlalchemy==2.0.28
alembic==1.13.1
stripe==8.5.0
